+++ www/jezf00/cv05/database/DatabaseOperations.php
@@ -1,0 +1,9 @@
+<?php
+interface DatabaseOperations
+{
+  public function fetch();
+  public function create($args);
+  public function save($args);
+  public function delete();
+}
+?>
