+++ www/jezf00/cv05/database/OrdersDB.php
@@ -1,4 +1,5 @@
 <?php
+require_once './database.php';
 class OrdersDB extends Database
 {
   public function create($args)
