+++ www/jezf00/cv05/database/database.php
@@ -1,11 +1,5 @@
 <?php
-interface DatabaseOperations
-{
-  public function fetch();
-  public function create($args);
-  public function save($args);
-  public function delete();
-}
+require_once './DatabaseOperations.php';
 
 abstract class Database implements DatabaseOperations
 {
@@ -29,30 +23,4 @@ abstract class Database implements DatabaseOperations
     return $this->separator;
   }
 }
-
-require './database/UsersDB.php'; 
-require './database/ProductsDB.php';
-require  './databese/OrdersDB.php';
-
-
-$users = new UsersDB();
-echo $users, PHP_EOL;
-$users->create(['name' => 'Filip', 'age' => 21]);
-$users->create(['name' => 'David', 'age' => 22]);
-$users->fetch();
-$users->delete();
-echo PHP_EOL;
-
-$products = new ProductsDB();
-echo $products, PHP_EOL;
-$products->create(['name' => 'Table', 'price' => 89]);
-$products->create(['name' => 'Chair', 'price' => 64]);
-$products->fetch();
-echo PHP_EOL;
-
-$orders = new OrdersDB();
-echo $orders, PHP_EOL;
-$orders->create(['number' => 15]);
-$orders->fetch();
-echo PHP_EOL;
 ?>
